video-service/requirements.txt
# KoboldAI Integration
httpx>=0.24.0
aiohttp>=3.8.0
websockets>=11.0

# Character Consistency & LoRA Training
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.7.0
safetensors>=0.3.0

# Video processing
moviepy>=1.0.3
opencv-python>=4.8.0
Pillow>=10.0.0

# Audio processing (existing TTS integration)
librosa>=0.10.0
soundfile>=0.12.0

# Image generation and processing
compel>=2.0.0
controlnet-aux>=0.0.6

# Core dependencies
torch>=2.4.1
torchaudio>=2.4.1
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.21.0

# API and async support
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Database connectivity (matching PHP backend)
mysql-connector-python>=8.2.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0

# Development tools
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0